Overwrite target file when reencoding
Previously, when files were reused for round trip reencoding, ``repack``
appended to the target file. This caused reencoded GRIBs to be compared with
messages from previous files when multiple input files were provided. Now the
target file is rewritten on the first message so that comparisons to verify
reencoding accuracy compare the same messages.